0.9.5 alpha release -Added portable scanner -Rewrote packets to not suck (as much) -Fixed multiple small graphical glitches and inconsistencies -Updated single block furnace textures -Rebalanced bench upgrade slots so tier is more relevant and quantity is less relevant to overall improvements -Changed Withered End structure block recipe to give 5 blocks instead of 2. -Fixed some display text for structure/multiblock control that was not showing up correctly.

+++ mal/api/IFuelContainer.java
@@ -18,14 +18,14 @@ public interface IFuelContainer {
 	 * @param fuelTime	the amount of fuel to add/remove
 	 * @param absolute	if this is true, the fuelTime value will be the new value, not added to the existing value
 	 */
-	public boolean setFuel(ItemStack is, int fuelTime, boolean absolute);
+	public boolean setFuel(ItemStack is, long fuelTime, boolean absolute);
 	
 	/**
 	 * 
 	 * @param is		the itemstack of the inventory item to interact with
 	 * @return			the amount of fuel in that itemstack
 	 */
-	public int getFuelValue(ItemStack is);
+	public long getFuelValue(ItemStack is);
 }
 
 /*******************************************************************************
